Moves MediaEditorOperation extension to it's own file
