+++ WordPress/Classes/ViewRelated/Aztec/Media/MediaEditorOperation+Description.swift
@@ -1,0 +1,18 @@
+import Foundation
+
+extension Array where Element == MediaEditorOperation {
+    var description: String {
+        return self.map { $0.description }.joined(separator: ", ")
+    }
+}
+
+extension MediaEditorOperation {
+    var description: String {
+        switch self {
+        case .crop:
+            return "crop"
+        case .rotate:
+            return "rotate"
+        }
+    }
+}
