+++ WordPress/Classes/ViewRelated/Aztec/ViewControllers/AztecPostViewController.swift
@@ -3462,20 +3462,3 @@ extension AztecPostViewController {
         attachment.uploadID = media.uploadID
     }
 }
-
-extension Array where Element == MediaEditorOperation {
-    var description: String {
-        return self.map { $0.description }.joined(separator: ", ")
-    }
-}
-
-extension MediaEditorOperation {
-    var description: String {
-        switch self {
-        case .crop:
-            return "crop"
-        case .rotate:
-            return "rotate"
-        }
-    }
-}
